FormStore helpers extracted and fixed. Source maps incorporated into webpack. (#58)
* Controlled mechanism implemented and onChange types for Field and Form updated.

* FormStore helpers extracted and fixed. Source maps incorporated into webpack.

FormStore helpers moved into a separate file and arrow function definition checking fixed in PropsEqual method. SourceMaps integrated into webpack pipeline.
