+++ packages/simplr-forms-dom/tools/webpack.config.ts
@@ -63,6 +63,16 @@ module.exports = {
     module: {
         rules: [
             {
+                enforce: "pre",
+                test: /\.js$/,
+                loader: "source-map-loader"
+            },
+            {
+                enforce: "pre",
+                test: /\.tsx?$/,
+                use: "source-map-loader"
+            },
+            {
                 test: /\.tsx?$/,
                 loader: "ts-loader",
                 options: {
@@ -73,5 +83,6 @@ module.exports = {
     resolve: {
         extensions: [".ts", ".tsx"]
     },
-    externals: externalsResolver
+    externals: externalsResolver,
+    devtool: "inline-source-map"
 };
