+++ packages/simplr-forms/package.json
@@ -45,6 +45,7 @@
     "react-test-renderer": "^15.5.4",
     "simplr-mvdir": "0.0.1-beta.7",
     "sinon": "^2.2.0",
+    "source-map-loader": "^0.2.1",
     "ts-jest": "^20.0.4",
     "ts-loader": "^2.1.0",
     "tslint": "^5.2.0",
