+++ packages/simplr-forms/src/stores/index.ts
@@ -1,2 +1,3 @@
 export * from "./form-store";
 export * from "./form-stores-handler";
+export * from "./form-store-helpers";
