+++ packages/simplr-forms/webpack.config.js
@@ -58,6 +58,16 @@ module.exports = {
     module: {
         rules: [
             {
+                enforce: "pre",
+                test: /\.js$/,
+                loader: "source-map-loader"
+            },
+            {
+                enforce: "pre",
+                test: /\.tsx?$/,
+                use: "source-map-loader"
+            },
+            {
                 test: /\.tsx?$/,
                 loader: "ts-loader",
                 options: {}
@@ -67,5 +77,6 @@ module.exports = {
     resolve: {
         extensions: [".ts", ".tsx"]
     },
-    externals: externalsResolver
+    externals: externalsResolver,
+    devtool: "inline-source-map"
 };
